Move VisualizationSettings from angular service to lib
