+++ frontend/src/dashboard/dashboard.module.js
@@ -3,8 +3,7 @@ var Dashboard = angular.module('metabase.dashboard', [
     'ngRoute',
     'metabase.directives',
     'metabase.services',
-    'metabase.dashboard.controllers',
-    'metabase.card.services'
+    'metabase.dashboard.controllers'
 ]);
 
 Dashboard.config(['$routeProvider', function ($routeProvider) {
